workflows: fix grobid tests  * ref https://github.com/cern-sis/issues-inspire/issues/1417
